--- a/bench/README.md
+++ b/bench/README.md
@@ -1,0 +1,4 @@
+This suite is taken with modifications from the Mithril example in
+[Merri/nom](https://github.com/Merri/nom/tree/gh-pages/todomvc/benchmark/mithril-new).
+
+(For example, it uses Mithril itself to help simplify the business logic.)