--- a/bench/app/.gitignore
+++ b/bench/app/.gitignore
@@ -1,0 +1,6 @@
+node_modules/todomvc-app-css/*
+!node_modules/todomvc-app-css/index.css
+
+node_modules/todomvc-common/*
+!node_modules/todomvc-common/base.js
+!node_modules/todomvc-common/base.css