… bench/app/index.html
… bench/app/readme.md
… bench/index.html
… bench/resources/tests.js
… package.json
… test/index.html
… test/mithril.js
… test/mithril.render.js
… test/mithril.trust.js