A summary of the craziness:

1. There now exists a benchmark suite. It's very heavily adapted from [Merri/nom's TodoMVC benchmarks](https://github.com/Merri/nom/tree/gh-pages/todomvc/benchmark), with the other versions eliminated and the benchmark suite using Mithril utilities to simplify some parts (e.g. ditching a home-grown deferred implementation).

2. Prototypes are avoided at all costs in the renderer. C-like structs are exclusively used outside of DOM and the API-related parts. This helps significantly in both speed and size. I also purposefully avoided using `this`, since that was actually slower.

3. The deferred implementation had a couple functions refactored into static equivalents. It also had some functions lifted out to the main constructor closure.

4. Only 1 test fails locally now.

5. Several names were changed to be much smaller and some of the exports were aliased. This was a pure size improvement for free.

6. The attribute and selector regexes are inlined to allow better engine optimization and caching. It takes a little less memory, and it's significantly faster, as engines can generate bytecode specific to the method used.

7. The version string was inlined. It's still at the top, though. A few bytes for free, and strings are cheap, anyways.

Overall, it's about a 25-35% (up to 40% in some cases) speed improvement on Chrome (and likely Edge, given similar internals) in the rendering for the general use case at the cost of ~200 bytes minified+gzipped. I hit that range consistently in the benchmark here. It tests multiple components, large node trees, controllers, etc., and it's complete with a small model.

The performance is also far more predictable. The benchmark at 500 todos is nearly always about 5000&pm;150ms on Chrome and with this patch, while I get anywhere from 6565ms to 8170ms on Chrome with the current `next`, generally around 6900&pm;400ms at 500 todos. (My computer is relatively weak - 4 GB of RAM on a 6-year-old consumer-grade laptop.)

On Firefox, the performance is roughly the same with or without this patch, running in the range of ~6400ms to ~6850ms, with equal consistency.

Or in other words, your app should fly (as if it weren't already :smile:).