@venning 2015-12-15
Is there any way to get GitHub to show the line diffs for `mithril.js`?  Right now it just shows "985 additions, 891 deletions not shown".