@isiahmeadows 2015-12-15
Also, this may create a few merge conflicts, because of the few fundamental algorithmic changes I made.