@isiahmeadows 2015-12-15
@venning Try https://github.com/lhorie/mithril.js/pull/887.diff. You won't have the helpful UI, but just search for `mithril.js`, and you should find it easily enough.