@isiahmeadows 2015-12-15
Or you can see that file pretty quickly by searching `2/5` in https://patch-diff.githubusercontent.com/raw/lhorie/mithril.js/pull/887.patch