@venning 2015-12-15
@isiahmeadows, thanks.  I wound up just checking out your repo and branch and running

    git diff --no-prefix -U9999 ..next mithril.js