@lhorie 2015-12-15
First thoughts:

1 - wow, this is a huge PR
2 - it's a little hard to understand the diff since a bunch of things got moved around
3 - benchmark looks like an awesome addition