@barneycarroll 2015-12-15
Amazing. Point 2 is very interesting, as is inlining the RegEx for JIT optimisation.

Adding benchmarks … brilliant.

Fantastic work you've been doing @impinball!