@isiahmeadows 2015-12-15
Nope ;)

On Tue, Dec 15, 2015, 16:02 Barney Carroll <notifications@github.com> wrote:

> @isiahmeadows <https://github.com/isiahmeadows> would it offend you if I
> told you I hadn't noticed? ;)
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/lhorie/mithril.js/pull/887#issuecomment-164895423>.
>