@isiahmeadows 2015-12-15
@lhorie Sorry about the large reorg. It's because I'm passing what are effectively structs around instead of 6-10 different variables simultaneously (very fragile and error-prone). Imagine OO-tendencies in C - static dispatch, but structs carry all the information, and there's no real inheritance whatsoever, just composition.