@isiahmeadows 2015-12-15
Here's my size comparison:

- `next`: 21.2 bytes minified, 8273 bytes minified + gzipped (default, 8265 bytes max)
- this: 22.1 bytes minified, 8497 bytes minified + gzipped (default, 8490 bytes max)

Yes, Mithril is starting to grow a little, but golfed code isn't usually fast.