@Bondifrench 2015-12-15
Nicely done @isiahmeadows 
It's very readable code.

I don't know the performance implications of this, but I'm wondering if at least for readability it would be worth refactoring these functions:
```
function nodeHasBody(node) {
	return node !== "AREA" &&
		node !== "BASE" &&
		node !== "BR" &&
		node !== "COL" &&
		node !== "COMMAND" &&
		node !== "EMBED" &&
		node !== "HR" &&
		node !== "IMG" &&
		node !== "INPUT" &&
		node !== "KEYGEN" &&
		node !== "LINK" &&
		node !== "META" &&
		node !== "PARAM" &&
		node !== "SOURCE" &&
		node !== "TRACK" &&
		node !== "WBR"
}
```
and
```
function shouldSetAttrDirectly(attr) {
	return attr !== "list" &&
		attr !== "style" &&
		attr !== "form" &&
		attr !== "type" &&
		attr !== "width" &&
		attr !== "height"
}
```
to something along the lines of:
```
var reservedNodes = ["AREA", "BASE", "BR", "COL", "COMMAND", "EMBED", "HR", "IMG", "INPUT", "KEYGEN", "LINK", "META", "PARAM", "SOURCE", "TRACK", "WBR"];
var reservedAttr = ["list", "style", "form", "type", "width", "height"];

function nodeHasBody (node) {
	return reservedNodes.every(function (i) {
		return i !== node;
	})
}

function shouldSetAttrDirectly(attr) {
	return reservedAttr.every(function (i) {
		return i !== attr;
	})
}
```
or maybe use the `forEach` function used at the top of the file.