@ornicar 2015-12-16
As the developer of lichess.org, that scores 45 million pageviews a month and uses mithril everywhere, I'm very glad to see this PR :) Performance is critical for us, since mithril powers the chess board interaction, both on the website and the mobile.

Thanks for the great work, @lhorie & @isiahmeadows !