@veloce 2015-12-16
As the dev of lichess mobile app I will follow @ornicar : thank you for this amazing work! As you know performance of VDOM is way more critical on mobile. I can't wait to see this PR merged!