@pygy 2015-12-16
re `reservedNodes`, what about a pseudo-set?

```JS
function setify(ary) {
    var res = {}
    ary.forEach(function(el){res[el]=true}
    return res
}
var emptyNodes = setify(["AREA", "BASE", "BR", "COL", "COMMAND", "EMBED", "HR", "IMG", "INPUT", "KEYGEN", "LINK", "META", "PARAM", "SOURCE", "TRACK", "WBR"])

var reservedAttr = setify(["list", "style", "form", "type", "width", "height"])
```