@isiahmeadows 2015-12-16
Actually, regular expressions appear to be faster, enough to make a small, but noticeable difference in benchmarks.